\name{read.CX}
\alias{read.CX}
\title{Read in a ClustalX .aln alignment file}
\description{Reads in a ClustalX .aln alignment file and parses it into a data frame.}
\usage{read.CX(file)}
\arguments{
  \item{file}{a connection or character string giving the name of the file to load.}
}
\Details{}
\Value{}
\Author{Kurt Wollenberg}
\References{}
\keyword{file}
