\name{read.FASTA}
\alias{read.FASTA}
\title{Read in a FASTA alignment file}
\description{Reads in a FASTA alignment file and parses it into a data frame. The FASTA file must be in .txt format.}
\usage{read.FASTA(file)}
\arguments{
  \item{file}{a connection or character string giving the name of the file to load.}
}
\Details{}
\Value{}
\Author{Kurt Wollenberg}
\References{}
\keyword{file}
