\name{read.GDoc}
\alias{read.GDoc}
\title{Read in a GeneDoc alignment file}
\description{Reads in a GeneDoc alignment file and parses it into a data frame. The GeneDoc file must be in .msf format.}
\usage{read.GDoc(file)}
\arguments{
  \item{file}{a connection or character string giving the name of the file to load.}
}
\Details{}
\Value{}
\Author{Kurt Wollenberg}
\References{}
\keyword{file}
